requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==4.0.1
django-admin-json-editor==0.2.3
django-cors-headers==3.10.1
django-environ==0.8.1
django-jsoneditor==0.2.2
django-jsonfield==1.4.1
djangorestframework==3.13.1
environ==1.0
idna==3.3
jsonfield==3.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
